server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-organizer-server"
version = "0.1.0"
description = "Phase 1 backend for a production-style photo organizer"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.116.0",
  "pillow>=11.3.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.10.0",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "pytest>=8.4.0",
  "pytest-cov>=6.2.0",
  "ruff>=0.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.lint.per-file-ignores]
"app/api/routers/*.py" = ["B008"]

[tool.setuptools.packages.find]
include = ["app*"]
